services/proxy/requirements.txt
Flask==3.0.3
gunicorn==23.0.0
Werkzeug==3.0.3
PyNaCl==1.5.0
google-cloud-pubsub==2.18.4
requests==2.31.0

